+++ src/main/java/com/xmu/cms/controller/UserController.java
@@ -1,12 +1,16 @@
 package com.xmu.cms.controller;
 
-import com.xmu.cms.entity.*;
+import com.xmu.cms.entity.Attendance;
+import com.xmu.cms.entity.Round;
+import com.xmu.cms.entity.Seminar;
+import com.xmu.cms.entity.Team;
 import com.xmu.cms.service.*;
 import com.xmu.cms.support.MyUser;
 import com.xmu.cms.support.UserInfo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
@@ -58,32 +62,32 @@ public class UserController {
 
 
     @GetMapping(value = "/course/{courseId}/rounds")
-    public List<Round> getRounds(@PathVariable("courseId") Integer courseId) {
+    public List<Round> getRounds(@PathVariable("courseId") BigInteger courseId) {
         return seminarService.getRoundsByCourseId(courseId);
     }
 
     @GetMapping(value = "/course/{courseId}/seminars")
-    public List<Seminar> getSeminarsByCourseId(@PathVariable("courseId") Integer courseId) {
+    public List<Seminar> getSeminarsByCourseId(@PathVariable("courseId") BigInteger courseId) {
         return seminarService.getSeminarsByCourseId(courseId);
     }
 
     @GetMapping(value = "round/{roundId}/seminars")
-    public List<Seminar> getAllSeminarInRound(@PathVariable(value = "roundId") Integer roundId) {
+    public List<Seminar> getAllSeminarInRound(@PathVariable(value = "roundId") BigInteger roundId) {
         return seminarService.getAllSeminarInRound(roundId);
     }
 
     @GetMapping(value = "/seminar/{seminarId}/teams")
-    public List<Team> getTeamsInSeminar(@PathVariable("seminarId") Integer seminarId) {
+    public List<Team> getTeamsInSeminar(@PathVariable("seminarId") BigInteger seminarId) {
         return teamService.getAllTeamsInSeminar(seminarId);
     }
 
     @GetMapping(value = "/seminar/{seminarId}")
-    public Seminar getSeminarById(@PathVariable("seminarId") Integer seminarId) {
+    public Seminar getSeminarById(@PathVariable("seminarId") BigInteger seminarId) {
         return seminarService.getSeminarBySeminarId(seminarId);
     }
 
     @GetMapping(value = "/seminar/{seminarId}/attendances")
-    public List<Attendance> getAttendancesInSeminar(@PathVariable("seminarId") Integer seminarId) {
+    public List<Attendance> getAttendancesInSeminar(@PathVariable("seminarId") BigInteger seminarId) {
         return seminarService.getAttendancesInSeminar(seminarId);
     }
